Update Texas Accidents.R
There are thousands of accidents every year, that all get recorded with different descriptors and data points. The data that I focused on was Accident severity, impact time, weather, location, and TMC code descriptions. 

My goal was to understand accident factors and provide relevant information that is applicable to governments to mitigate these risks and to better direct safety efforts. 

The original severity measure was not very informative, with all the accidents fitting in two of the measures. I opened the severity measures based on how long the accident took to rectify. This new measurement broadened the range of the severity measure and allowed me to get more accurate results. 

The results were very interesting, being that there were more accidents while it was sunny outside, compared to rainy or cloudy. I figure this is because people are more relaxed and pay attention to the road less while it is a nice day outside, compared to heavy rain. 

Text mining the TMC code descriptions allowed me to see that the most common reason for an accident is either a blocked right lane or two blocked lanes. The words accident occurred in every result, so I had to use stop words to ensure that they didn't take over the word cloud or my frequency measures. 

Using a decision tree, I found that weekdays increase the risk of a severe accident, with a 74% accuracy. 

Lastly, I used a Logistic Regression Model to find the ratio of accidents compared to the different determinants.