Add files via upload
This project was to analyze the text of various documents that had to do with ebola, and the scientific findings that the researchers discovered.

First, I started by making a corpus and creating a document term matrix utilizing that corpus. Using the document term matrix, I was able to create a list of the most frequent terms, using the "freq.sorted" function.

Using a word cloud, I was able to not only document the most frequent words, but was able to provide a simple graphic that was very easy to understand. This graphic allowed me to visualize the differences in the frequency of each word.

Lastly, I used a LDA model to see which words corresponded with each topic. I generated 12 different topics and was able to see which words were being used with each other at a high rate and the level of correspondence from 0-1. Because there were 14 documents, I had to breakdown which topics were most valuable to each of the documents. Using this, I could have a general understanding of what the document was talking about without even reading the whole thing.