Add files via upload
Because I want to get into the marketing field, I decided to do a project surrounding predictive modeling for marketing purposes. So, I downloaded data from a bank, wanting to devise a classification model and predict whether or not a potential customer is a likely buyer before deciding whom to contact. Therefore, my job was to help the firm to build a predictive model using historical data with 4,521 observations.

To start I created a train variable, something that could build the model using 2/3 of the data, and created a test variable that would test and calibrate the model. To build the model, I used R part and a minsplit of 50. I then built a decision tree usins ghte data from the model.

Then using a confusion matrix, I was able to complete my final prediction. I computed the performance measures to test the false positivity rate, false-negative rate, specificity, and sensitivity. I also had to calculate the accuracy of the model. The accuracy I calculated was .88, which is high but not perfect. I could adjust some of the variables earlier on to ensure that that it has at least 90% accuracy.

But with this level of accuracy, I would be able to make informed decisions to the bank with accurate results on who is more likely to buy from the bank.